+++ src/anyio/__init__.py
@@ -18,6 +18,8 @@ from ._core._exceptions import ClosedResourceError as ClosedResourceError
 from ._core._exceptions import ConnectionFailed as ConnectionFailed
 from ._core._exceptions import DelimiterNotFound as DelimiterNotFound
 from ._core._exceptions import EndOfStream as EndOfStream
+from ._core._exceptions import FutureAlreadyFinished as FutureAlreadyFinished
+from ._core._exceptions import FutureCancelled as FutureCancelled
 from ._core._exceptions import IncompleteRead as IncompleteRead
 from ._core._exceptions import NoEventLoopError as NoEventLoopError
 from ._core._exceptions import RunFinishedError as RunFinishedError
@@ -30,6 +32,7 @@ from ._core._fileio import AsyncFile as AsyncFile
 from ._core._fileio import Path as Path
 from ._core._fileio import open_file as open_file
 from ._core._fileio import wrap_file as wrap_file
+from ._core._futures import Future as Future
 from ._core._resources import aclose_forcefully as aclose_forcefully
 from ._core._signals import open_signal_receiver as open_signal_receiver
 from ._core._sockets import TCPConnectable as TCPConnectable