+++ src/anyio/_core/_exceptions.py
@@ -175,3 +175,17 @@ class TaskNotFinished(Exception):
     Raised when attempting to access the return value or exception of a
     :class:`.TaskHandle` that is still pending completion.
     """
+
+
+class FutureCancelled(Exception):
+    """
+    Raised when attempting to access the return value or exception of a
+    :class:`.Future` that was cancelled.
+    """
+
+
+class FutureAlreadyFinished(Exception):
+    """
+    Raised when attempting set a result of or await a
+    :class:`.Future` that has already completed.
+    """